% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdeSetupFunctions.R
\name{indicator}
\alias{indicator}
\title{The indicator function of an event}
\usage{
indicator(bool)
}
\arguments{
\item{bool}{a boolean for some event, statement, condition}
}
\value{
1 if true, zero if false
}
\description{
{A wrapper to \code{ifelse(bool, 1, 0)} to represent the indicator
of an event 'bool'.}
}
\details{
{Used like \code{indicator(x == 0)}, etc in probability theory.}
}
