% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdeSetupFunctions.R
\name{setupPDE}
\alias{setupPDE}
\title{Initialize grids for the PDE solver}
\usage{
setupPDE(dynamics, problem, region, N, M)
}
\arguments{
\item{dynamics}{list of functions defining the dynamics of the Ito process}

\item{problem}{list of functions defining the Feynman-Kac problem}

\item{region}{vector of maturity, lower and upper bounds of space}

\item{N}{time resolution}

\item{M}{space resolution}
}
\value{
list
}
\description{
{Initializes grids used in the Feyman-Kac PDE solver.}
}
\details{
{The argument \code{dynamics} must be a named list of functions
\itemize{
\item \code{drift}, a function of \eqn{(t, x)} representing the local drift
\item \code{volat}, a function of \eqn{(t, x)} representing the local volatility
}
while the argument \code{problem} must be a named list of functions
\itemize{
\item \code{discount}, a function of \eqn{(t, x)} representing the discount function
\item \code{running_cost}, a function of \eqn{(t, x)} representing the running cost
\item \code{terminal_cost}, a function of \eqn{(x)} representing the terminal cost, as only a function of \eqn{x}.
}
and finally the argument \code{region} must be a vector with entries representing the maturity,
lower bound, upper bound, and time and space resolutions, in that order.}
}
