% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutionAnimator.R
\name{solutionAnimator}
\alias{solutionAnimator}
\title{Animate the solution to a Feynman-Kac problem over time}
\usage{
solutionAnimator(fk, backwards = TRUE, detail = 10)
}
\arguments{
\item{fk}{the solution grid returned from \code{solvePDE}}

\item{backwards}{boolean for terminal condition (TRUE) or initial condition (FALSE)}

\item{detail}{detail in \code{animate} from \code{gganimate}}
}
\value{
null
}
\description{
{Animates the time evolution of \eqn{u(t,x)} of the conditional expectation of a
function of a diffusion given its starting point.}
}
