% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdeSetupFunctions.R
\name{solvePDE}
\alias{solvePDE}
\title{Compute conditional expectation via solving Feynman-Kac PDE}
\usage{
solvePDE(
  dynamics = NULL,
  problem = NULL,
  region = NULL,
  control = list(N = 100, M = 100, variational = TRUE, output = "greeks", engine =
    "c++")
)
}
\arguments{
\item{dynamics}{list of functions defining the dynamics of the Ito process , see details}

\item{problem}{list of functions defining the Feynman-Kac problem, the conditional expectation}

\item{region}{vector, the space-time region to solve over and resolution of grids, see details}

\item{control}{list of time and space resolutions, variational boolean, and output option, and engine (r or c++)}
}
\value{
list or numeric
}
\description{
{Compute conditional expectations of functions of diffusions/Ito processes
via solving the Feynman-Kac PDE.}
}
\details{
{The argument \code{dynamics} must be a named list of functions
\itemize{
\item \code{drift}, a function of \eqn{(t, x)} representing the local drift
\item \code{volat}, a function of \eqn{(t, x)} representing the local volatility
}
while the arguments of \code{problem} must be a named list of functions
\itemize{
\item \code{discount}, a function of \eqn{(t, x)} representing the discount function
\item \code{running_cost}, a function of \eqn{(t, x)} representing the running cost
\item \code{terminal_cost}, a function of \eqn{(x)} representing the terminal cost, as only a function of \eqn{x}.
}
and finally the argument \code{region} must be a vector with entries representing the maturity,
lower bound, upper bound.}
}
