% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tridiagSolver.R
\name{thomas_algorithm}
\alias{thomas_algorithm}
\title{Solving tridiagonal linear systems fast}
\usage{
thomas_algorithm(alpha, beta, delta, b.vec)
}
\arguments{
\item{alpha}{lower diagonal}

\item{beta}{diagonal}

\item{delta}{upper diagonal}

\item{b.vec}{the auxillary vectory in "Ax=b"}
}
\value{
vector
}
\description{
{Thomas algorithm for solving tridiagonal systems fast without inverting any matrices.}
}
\details{
{The algorithm is fairly well known, i.e. on wikipedia and not complicated to implement. The tridiagonal system allows one to use back-substitution.}
}
